refactor: 피드백 1차 반영
- Graph에서 Station을 통한 커뮤니케이션으로 Path 생성, stations 멤버 변수 삭제
- Path 단위 테스트 추가(duration 계산 기능 확인)
- Graph 단위 테스트 추가(주어진 기준에 따라 path를 생성하는지 확인)
- Controller/Service의 역할이 드러나도록 패키지 네이밍 수정 (client -> map)
- Map관련 요청 url 수정
- PathResponse 불필요한 생성자 삭제, 멤버 변수 타입 수정

+++ src/main/java/wooteco/subway/domain/path/StationWeightEdge.java
@@ -5,10 +5,10 @@ import org.jgrapht.graph.DefaultWeightedEdge;
 import wooteco.subway.domain.LineStation;
 
 public class StationWeightEdge extends DefaultWeightedEdge {
-    private double distance;
-    private double duration;
+    private int distance;
+    private int duration;
 
-    public StationWeightEdge(double distance, double duration) {
+    public StationWeightEdge(int distance, int duration) {
         this.distance = distance;
         this.duration = duration;
     }
